apps/agents/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.0.0
requests==2.31.0
aiohttp==3.9.0
httpx==0.25.0
pytest==7.4.0
pytest-asyncio==0.21.0

